requirements.txt
fastapi~=0.104.1
pydantic~=2.4.2
pycryptodome~=3.19.0
uvicorn~=0.24.0.post1